Backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-dotenv
langchain-community
langchain-qdrant
sentence-transformers
qdrant-client
google-generativeai
openai
